define stricter type

+++ header_and_config/LUR7_can.c
@@ -388,7 +388,7 @@ void _can_handle_RXOK() {
 	uint8_t data[dlc]; // create vector for data
 
 	//read data
-	for (int i = dlc; i > 0; i--) {
+	for (uint8_t i = dlc; i > 0; i--) {
 		data[i-1] = CANMSG; //CANMSG autoincrements, !AINC = 0. has to be reversed to send data non inverted. No idea why, stupid design of CAN controller
 	}
 
@@ -411,7 +411,7 @@ void _can_handle_TXOK() {
 	uint8_t data[dlc]; // create vector for data
 
 	//read data
-	for (int i = dlc; i > 0; i--) {
+	for (uint8_t i = dlc; i > 0; i--) {
 		data[i-1] = CANMSG; //CANMSG autoincrements, !AINC = 0. has to be reversed to send data non inverted. No idea why, stupid design of CAN controller
 	}
